DVBSub2 support removed from TsReader.ax, and dshowhelper log rotation changed to daily (to match TsReader.ax). Also some compile warnings fixed in TsReader.ax

+++ DirectShowFilters/TsReader/source/MpegPesParser.cpp
@@ -56,7 +56,7 @@ bool CMpegPesParser::ParseVideo(byte* tsPacket,bool isMpeg2,bool reset)
 	if (isMpeg2)
 	{
 		seqhdr seq;
-		if (hdrParser.Read(seq,framesize,&pmt,reset))
+		if (hdrParser.Read(seq,(int)framesize,&pmt,reset))
 		{
 			//hdrParser.DumpSequenceHeader(seq);
 			basicVideoInfo.width=seq.width;
@@ -75,7 +75,7 @@ bool CMpegPesParser::ParseVideo(byte* tsPacket,bool isMpeg2,bool reset)
 	}
 	else 
 	{
-		if (hdrParser.Read(avc,framesize,&pmt,reset))
+		if (hdrParser.Read(avc, (int)framesize,&pmt,reset))
 		{
 			//hdrParser.DumpAvcHeader(avc);
 			basicVideoInfo.width=avc.width;
